+++ routes/web.php
@@ -19,47 +19,54 @@ use App\Http\Controllers\ProductController;
 */
 //list all products
 Route::get('/', [ProductController::class, 'index']);
+
 //show single product
 Route::get('/product/{product}', [ProductController::class, 'show']);
 
 //show create form
-Route::get('/products/create', [ProductController::class, 'create'])->name('products.create');
+Route::get('/products/create', [ProductController::class, 'create'])->name('products.create')->middleware('auth');
 
 //save form
-Route::post('/product', [ProductController::class, 'store']);
+Route::post('/product', [ProductController::class, 'store'])->middleware('auth');
 
 //delete record
-Route::delete('/products/{product}', [ProductController::class, 'destroy']);
+Route::delete('/products/{product}', [ProductController::class, 'destroy'])->middleware('auth');
 
 //update form
-Route::get('/product/{product}/edit', [ProductController::class, 'edit']);
+Route::get('/product/{product}/edit', [ProductController::class, 'edit'])->middleware('auth');
 
 //save update form
-Route::put('/products/{product}', [ProductController::class, 'update']);
+Route::put('/products/{product}', [ProductController::class, 'update'])->middleware('auth');
 
 //register here
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 //save registration
-Route::post('/users', [UserController::class, 'store']);
+Route::post('/users', [UserController::class, 'store'])->middleware('guest');
 
 //logout
 Route::post('/logout', [UserController::class, 'logout']);
 
 //show login form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 //autenticate users login
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
 
 //manage products
-Route::get('/products/manage', [ProductController::class, 'manage'])->name('products.manage');
+Route::get('/products/manage', [ProductController::class, 'manage'])->name('products.manage')->middleware('auth');
 
 //cart components
-Route::get('/cart', [CartController::class,'shoppingCart']);
+Route::get('/cart', [CartController::class, 'shoppingCart'])->name('products.cart')->middleware('auth');
 
-Route::delete('/cart/{product}/remove',[CartController::class, 'remove']);
+//cart remove item
+Route::delete('/cart/{product}/remove', [CartController::class, 'remove'])->middleware('auth');
 
+//cart get
 Route::get('/cart/{product}', [CartController::class, 'ProductCart']);
 
-Route::delete('/cart/{product}/destroyall', [CartController::class, 'destroyall']);
+//cart destroy everything
+Route::delete('/cart/{product}/destroyall', [CartController::class, 'destroyall'])->middleware('auth');
+
+//addtocart
+// Route::post('/cart/{id}', 'ProductController@addProducttoCart')->name('addToCart');
